Issue #17 experiment with Groq API
The Groq SDK appears to require a backend to function.  I haven't gotten
anything in the groq-api.js file to work yet.
I had some console logs set up in App.jsx to test things but I have
since deleted them.

I reached out to the Groq team and submitted a support ticket on the
Chingu Discord to see if I overlooked a way to send requests to
the Groq API with frontend only.  The docs do not have info on
URL endpoints to use with fetch.