Issue #17 - set up AI response state in App.jsx
Imported useState. Added aiResponse state variable.  Also set testColor
state variable which is temporary.  testColor state passed to PromptForm
as the color prop. setAiResponse is also passed to PromptForm so that
on form submit the aiResponse state can be set to the response from
the Groq API. Test code added to App.jsx which console logs aiResponse if
the variable is truthy.

setAiResponse added as prop to PromptForm.jsx. handleFormSubmit refactored
to check whether contextValue is truthy (this field is required in the form).
getGroqChatCompletion is called on submit and setAiResponse sets the groqResponse
to the state variable in App.jsx.