Issue #19 - Add formatResponseText function
Added formatResponseText function in ResponseDisplay.jsx that takes
the AI response string, splits it on the \n characters, and maps over
the array with a break tag for each "" and a paragraph tag for each
non-empty string.  This allows the response to render in a much more
readable format.

Also added a response header that renders above the response for
clarity.  The response header was given a margin bottom of 1em in
ResponseDisplay.css.

Finally, the test code that console logged the AI response was deleted
from App.jsx since the response is now rendered on the page.