Issue #19 - Add component that displays color swatches
Created ColorDisplay.jsx, which takes color and colorRecs as props.
Displays a color swatch with associated color code for the selected color.
Then if colorRecs is truthy, it maps over colorRecs and creates swatches
for each of those colors as well.

ResponseDisplay.jsx imports ColorDisplay and returns it if color is
truthy (in other words, if a color has been selected).

ResponseDisplay.css updated with basic styling for color swatches and
their associated container divs.