Issue #57 - Refactor color code format state variables
Removed the state variable for color code format from PromptForm.jsx.
State variable in App.jsx (colorCodeFormat) passed to PromptForm as a prop
and set directly in the onChange function for the color code format dropdown.
This change was made because when I added the state variable to App.jsx, I had
unknowingly created two different state variables in different components that
did the exact same thing.

I also did some minor refactoring in ResponseDisplay.jsx. I moved hexRegex out
of extractRecommendedColors function and added a colorRegex parameter that
can be passed when the function is called. I also added colorCodeFormat as a
prop in this component.  Will need to write a function that returns the
appropriate regex depending on colorCodeFormat. Currently this component
can ONLY handle hex colors.

App.jsx is still passing the hex state variable to appropriate props
because of the above limitation, even though the hex state does not update
when the color wheel is used.