Issue #71 - [WIP, Optional] - start updating section headers
Started refactoring the header to MoreInfo.jsx to utilize the SectionHeader
component Tan made. However, this broke the layout more than I expected
due to differences in css between MoreInfo and TeamSection, but also the
SectionHeader styling doesn't work well with longer titles.

I'm putting this issue on hold, as it's an optional improvement and
there is other work that needs to be done that is required before
finalizing the project today (as it is due).

If desired, this can be picked back up as a future stretch goal.